requirements.txt
celery==4.1.0
colorama==0.3.9
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
ipython==6.2.1
marshmallow==2.14.0
mock==2.0.0
multidict==4.1.0
PyJWT==1.5.3
PyYAML==3.12
requests==2.18.4
SQLAlchemy==1.1.14

